remove executor add docstrings to all interfaces
