+++ src/main/java/ru/bomzheg/dispatcher/Filter.java
@@ -2,6 +2,14 @@ package ru.bomzheg.dispatcher;
 
 import org.telegram.telegrambots.meta.api.interfaces.BotApiObject;
 
+/**
+ * Filter is a class that receive in check param event and checks if this event is suitable
+ */
 public interface Filter {
+    /**
+     * receive event and make some checks
+     * @param event event need to be checked
+     * @return whether the event passed through the filter successfully
+     */
     boolean check (BotApiObject event);
 }
