+++ src/main/java/ru/bomzheg/dispatcher/SendHelper.java
@@ -3,6 +3,16 @@ package ru.bomzheg.dispatcher;
 import org.telegram.telegrambots.meta.api.objects.Message;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 
+/**
+ * helper class to easy send messages
+ */
 public interface SendHelper {
+    /**
+     * easy send message
+     * @param chatId where to send
+     * @param text what to send
+     * @return message on success
+     * @throws TelegramApiException on failure error from telegram
+     */
     Message sendMessage(long chatId, String text) throws TelegramApiException;
 }
